Added delimiter to boolean addition and intersection
Boolean operations now have an optional delimiter which will separate
each “set” (heat map data set). In the case of addition it’s the union
symbol ( ∪ ) and in the case of intersect it’s the intersection symbol
( ∩ ).

In addition, some basic code for intersection was added, but does not
work as intended.